+++ tulip/shared/tsequencer.h
@@ -19,7 +19,10 @@ extern uint32_t sequencer_period[SEQUENCER_SLOTS];
 extern uint32_t sequencer_tick[SEQUENCER_SLOTS];
 
 extern mp_obj_t defer_callbacks[DEFER_SLOTS];
-extern uint32_t defer_sysclock[DEFER_SLOTS];
+// 64-bit absolute deadline (get_time_ms()). This was uint32_t off the 32-bit
+// tick, so a defer straddling the 49.7-day rollover either fired instantly or
+// waited another 49.7 days.
+extern int64_t defer_sysclock[DEFER_SLOTS];
 extern mp_obj_t defer_args[DEFER_SLOTS];
